+++ apps/studio/src/lib/editor/engine/history/index.ts
@@ -1,101 +1,9 @@
 import { sendAnalytics } from '@/lib/utils';
-import type {
-    Action,
-    Change,
-    IndexActionLocation,
-    UpdateStyleAction,
-    WriteCodeAction,
-} from '@onlook/models/actions';
+import type { Action } from '@onlook/models/actions';
 import { jsonClone } from '@onlook/utility';
 import { makeAutoObservable } from 'mobx';
 import type { EditorEngine } from '..';
-import { assertNever } from '/common/helpers';
-
-function reverse<T>(change: Change<T>): Change<T> {
-    return { updated: change.original, original: change.updated };
-}
-
-function reverseMoveLocation(location: IndexActionLocation): IndexActionLocation {
-    return {
-        ...location,
-        index: location.originalIndex,
-        originalIndex: location.index,
-    };
-}
-
-function reverseStyleAction(action: UpdateStyleAction): UpdateStyleAction {
-    return {
-        ...action,
-        targets: action.targets.map((target) => ({
-            ...target,
-            change: reverse(target.change),
-        })),
-    };
-}
-
-function reverseWriteCodeAction(action: WriteCodeAction): WriteCodeAction {
-    return {
-        ...action,
-        diffs: action.diffs.map((diff) => ({
-            ...diff,
-            original: diff.generated,
-            generated: diff.original,
-        })),
-    };
-}
-
-function undoAction(action: Action): Action {
-    switch (action.type) {
-        case 'update-style':
-            return reverseStyleAction(action);
-        case 'insert-element':
-            return {
-                ...action,
-                type: 'remove-element',
-            };
-        case 'remove-element':
-            return {
-                ...action,
-                type: 'insert-element',
-                editText: null,
-            };
-        case 'move-element':
-            return {
-                ...action,
-                location: reverseMoveLocation(action.location),
-            };
-        case 'edit-text':
-            return {
-                ...action,
-                originalContent: action.newContent,
-                newContent: action.originalContent,
-            };
-        case 'group-elements':
-            return {
-                ...action,
-                type: 'ungroup-elements',
-            };
-        case 'ungroup-elements':
-            return {
-                ...action,
-                type: 'group-elements',
-            };
-        case 'write-code':
-            return reverseWriteCodeAction(action);
-        case 'insert-image':
-            return {
-                ...action,
-                type: 'remove-image',
-            };
-        case 'remove-image':
-            return {
-                ...action,
-                type: 'insert-image',
-            };
-        default:
-            assertNever(action);
-    }
-}
+import { undoAction, updateTransactionActions } from './helpers';
 
 enum TransactionType {
     IN_TRANSACTION = 'in-transaction',
@@ -160,14 +68,10 @@ export class HistoryManager {
 
     push = (action: Action) => {
         if (this.inTransaction.type === TransactionType.IN_TRANSACTION) {
-            // Only allow one action per type, otherwise, overwrite the existing action
-            if (this.inTransaction.actions.some((a) => a.type === action.type)) {
-                this.inTransaction.actions = this.inTransaction.actions.map((a) =>
-                    a.type === action.type ? action : a,
-                );
-                return;
-            }
-            this.inTransaction.actions.push(action);
+            this.inTransaction.actions = updateTransactionActions(
+                this.inTransaction.actions,
+                action,
+            );
             return;
         }
